book_logistics_data_lab/requirements.txt
pandas>=2.0
numpy>=1.24
psycopg2-binary>=2.9
pyarrow>=14.0
